+++ internal/cgroup/cgroup_windows.go
@@ -1,5 +1,11 @@
 package cgroup
 
-func start(c *Cgroup) {
+type Cgroup struct {
+	opt *CgroupOptions
+
+	err error
+}
+
+func (c *Cgroup) start() {
 	l.Infof("not support windows system, exit")
 }