+++ plugins/inputs/dialtesting/input.go
@@ -25,6 +25,8 @@ import (
 	dt "gitlab.jiagouyun.com/cloudcare-tools/kodo/dialtesting"
 )
 
+var _ inputs.ReadEnv = (*Input)(nil)
+
 var (
 	AuthorizationType = `DIAL_TESTING`
 	SignHeaders       = []string{
@@ -540,6 +542,29 @@ func (d *Input) pullHTTPTask(reqURL *url.URL, sinceUs int64) ([]byte, int, error
 			d.stopAlltask()
 		}
 		return nil, resp.StatusCode / 100, fmt.Errorf("pull task failed")
+	}
+}
+
+// ReadEnv support envs:
+// ENV_INPUT_DIALTESTING_AK: string
+// ENV_INPUT_DIALTESTING_SK: string
+// ENV_INPUT_DIALTESTING_REGION_ID: string
+// ENV_INPUT_DIALTESTING_SERVER: string.
+func (d *Input) ReadEnv(envs map[string]string) {
+	if ak, ok := envs["ENV_INPUT_DIALTESTING_AK"]; ok {
+		d.AK = ak
+	}
+
+	if sk, ok := envs["ENV_INPUT_DIALTESTING_SK"]; ok {
+		d.SK = sk
+	}
+
+	if regionID, ok := envs["ENV_INPUT_DIALTESTING_REGION_ID"]; ok {
+		d.RegionID = regionID
+	}
+
+	if server, ok := envs["ENV_INPUT_DIALTESTING_SERVER"]; ok {
+		d.Server = server
 	}
 }